// src/shared/api/data-contracts.ts
export interface UserResponse {
  /** User id */
  id: number;
  /** First name */
  first_name: string;
  /** Second name */
  second_name: string;
  /** Display name */
  display_name: string;
  /** User login - unique */
  login: string;
  /**
   * User email - unique
   * @format email
   */
  email: string;
  /** User phone */
  phone: string;
  /** Avatar */
  avatar: string;
}
// ...
export interface CreateChatRequest {
  /** Chat title */
  title: string;
}

export interface UsersRequest {
  users: number[];
  /** Chat id */
  chatId: number;
}

/** @example {"id":123,"title":"my-chat","avatar":"/123/avatar1.jpg","unread_count":15,"created_by":12345,"last_message":{"user":{"first_name":"Petya","second_name":"Pupkin","avatar":"/path/to/avatar.jpg","email":"my@email.com","login":"userLogin","phone":"8(911)-222-33-22"},"time":"2020-01-02T14:22:22.000Z","content":"this is message content"}} */
export interface ChatsResponse {
  /** Chat id */
  id: number;
  /** Chat title */
  title: string;
  /** Chat avatar */
  avatar: string;
  /** Number of unread messages in chat for current user */
  unread_count: number;
  last_message: {
    /**
     * Message user (sender)
     * @format date-time
     */
    user?: UserResponse;
    /**
     * Message timestamp
     * @format timestamp
     */
    time?: string;
    /** Message content */
    content?: string;
  };
}

export interface ChatDeleteRequest {
  /** Chat id */
  chatId: number;
}

/** @example {"userId":12,"result":{"id":123,"title":"deleted-chat","avatar":"/123/avatar1.jpg","created_by":12345}} */
export interface ChatDeleteResponse {
  /** User id */
  userId: number;
  result: ChatsResponse;
}

export interface ChatArchiveRequest {
  /** Chat id */
  chatId: number;
}

export interface ChatArchiveResponse {
  /** User id */
  userId: number;
  result: ChatsResponse;
}

export interface ChatsMessagesTokenResponse {
  /** Token for web socket server */
  token: string;
}

/** @example {"unread_count":12} */
export interface UnreadCountResponse {
  /** New messages count */
  unread_count: number;
}
// ...
export interface HttpErrorBody {
  /** Error message */
  reason: string;
}
// ...
export interface Resource {
  /** Message id */
  id: number;
  /** User id */
  user_id: number;
  /** Server relative file path */
  path: string;
  /** Initial file name */
  filename: string;
  /** File content type (e.g "image/jpeg" for .jpg images) */
  content_type: string;
  /** File size in bytes */
  content_size: number;
  /**
   * Resource uploading time
   * @format date-time
   */
  upload_date: string;
}

/** @example {"id":123,"user_id":231,"chat_id":312,"time":"2020-01-02T14:22:22.000Z","type":"file","content":132,"file":{"id":132,"user_id":231,"path":"/32543654dsf/434534r3rsddfs_my-file.jpg","filename":"my-file.jpg","content_type":"image/jpeg","content_size":543672,"upload_date":"2020-01-02T14:22:22.000Z"}} */
export interface ChatMessage {
  /** Message id */
  id: number;
  /** User id */
  user_id: number;
  /** Chat id */
  chat_id: number;
  /**
   * Message sent time
   * @format date-time
   */
  time: string;
  /** Message type */
  type: 'message' | 'file';
  /** Message content (message text for messages and resourceId for files) */
  content: string;
  /** File */
  file?: Resource;
}

/** @example {"id":123,"first_name":"petya","second_name":"petrov","display_name":"petya petrov","login":"my-login","avatar":"/path/to/my-file.jpg","role":"admin"} */
export interface ChatUserResponse {
  /** User id */
  id: number;
  /** First name */
  first_name: string;
  /** Second name */
  second_name: string;
  /** Display name */
  display_name: string;
  /** User login - unique */
  login: string;
  /** Avatar */
  avatar: string;
  /** User role */
  role: 'admin' | 'regular';
}